Renamed InitInput to InitParams, InitParams to ConstructorParams for clarity - renamed functions to get defaults accordingly and moved to utils.ts, Exposed required types from main file
